Enter today's wod and display it
Why:

Users (me) would like to see what the workout of the day is.

This PR:

Adds the wods model for storing the workout of the day. This has a date
which is the date of entry and a description. I've also added a unique
index to the date to ensure there is only ever 1 wod for each day.